feat(security): moves userauth, keyinfo, and error types to slim package
