+++ packages/client/package.json
@@ -35,12 +35,16 @@
     },
     {
       "path": "../context"
+    },
+    {
+      "path": "../security"
     }
   ],
   "dependencies": {
     "@improbable-eng/grpc-web": "^0.12.0",
     "@textile/context": "^0.4.3",
     "@textile/multiaddr": "^0.0.12",
+    "@textile/security": "^0.1.0",
     "@textile/threads-client-grpc": "^0.1.19",
     "@textile/threads-core": "^0.1.15",
     "@textile/threads-id": "^0.1.2",
