+++ packages/context/src/index.ts
@@ -1,4 +1,5 @@
 import { grpc } from '@improbable-eng/grpc-web'
+import { UserAuth, KeyInfo, expirationError } from '@textile/security'
 import { HMAC } from 'fast-sha256'
 import multibase from 'multibase'
 
@@ -11,50 +12,6 @@ export type HostString =
   | 'http://127.0.0.1:3007'
   | string
 export const defaultHost: HostString = 'https://api.textile.io:3447'
-
-/**
- * Type describing the required shape of input user group/account keys.
- */
-export type KeyInfo = {
-  /**
-   * API key. Can be embedded/shared within an app.
-   */
-  key: string
-  /**
-   * User group/account secret. Should not be embedded/shared publicly.
-   */
-  secret: string
-  /**
-   * Key type. One of ACCOUNT or USER
-   */
-  type: 0 | 1
-}
-
-/**
- * Type describing the minimal requirements for a user group auth session.
- */
-export type UserAuth = {
-  /**
-   * The public component of the user group key.
-   */
-  key: string
-  /**
-   * The signature of the authentication message.
-   */
-  sig: string
-  /**
-   * The authentication message.
-   */
-  msg: string
-  /**
-   * An optional authentication token as generated by getToken or getTokenChallenge.
-   */
-  token?: string
-}
-
-export const expirationError = new Error(
-  'Context expired. Consider calling withUserKey or withAPISig to refresh.',
-)
 
 /**
  * Generate an authorization signature and message.
