+++ packages/database/src/db.ts
@@ -1,7 +1,8 @@
 import toJsonSchema from 'to-json-schema'
 import cbor from 'cbor-sync'
 import { Network, Client } from '@textile/threads-network'
-import { Context, UserAuth, defaultHost } from '@textile/context'
+import { Context, defaultHost } from '@textile/context'
+import { UserAuth } from '@textile/security'
 import { EventEmitter2 } from 'eventemitter2'
 import { Dispatcher, Instance, DomainDatastore, Event, Update, Op } from '@textile/threads-store'
 import { Datastore, Key } from 'interface-datastore'
