+++ packages/security/README.md
@@ -1,0 +1,9 @@
+# Security _(security)_
+
+Common types/interfaces for Textile security including authentication and authorization.
+
+[![Textile](https://img.shields.io/badge/made%20by-Textile-informational.svg)](https://textile.io)
+[![License](https://img.shields.io/github/license/textileio/js-threads.svg)](../../LICENSE)
+[![npm (scoped)](https://img.shields.io/npm/v/@textile/security.svg)](https://www.npmjs.com/package/@textile/security)
+
+This sub-package is part of [`js-threads`](https://github.com/textileio/js-threads). See the [top-level documentation](https://textileio.github.io/js-threads) for details.
