+++ packages/security/lib/security.js
@@ -1,0 +1,7 @@
+'use strict';
+
+module.exports = security;
+
+function security() {
+    // TODO
+}
