+++ packages/security/package.json
@@ -1,6 +1,7 @@
 {
-  "name": "@textile/context",
-  "version": "0.4.3",
+  "name": "@textile/security",
+  "version": "0.1.0",
+  "description": "Common types/interfaces for Textile security including authentication and authorization.",
   "main": "dist/index",
   "types": "dist/index",
   "files": [
@@ -19,8 +20,7 @@
     "build:node": "npm run compile",
     "build:browser": "webpack --mode production --config ../../webpack.config.js",
     "compile": "tsc -b tsconfig.json",
-    "clean": "rimraf ./dist ./tsconfig.tsbuildinfo",
-    "test:browser": "polendina --cleanup --timeout 60 src/**/*.spec.ts --webpack-config ../../webpack.test.js"
+    "clean": "rimraf ./dist ./tsconfig.tsbuildinfo"
   },
   "repository": "github:textileio/js-threads",
   "devDependencies": {
@@ -28,19 +28,6 @@
     "rimraf": "^3.0.0",
     "typescript": "^3.7.2"
   },
-  "references": [
-    {
-      "path": "../id"
-    },
-    {
-      "path": "../security"
-    }
-  ],
   "dependencies": {
-    "@textile/security": "^0.1.0",
-    "@types/multibase": "^0.6.0",
-    "fast-sha256": "^1.3.0",
-    "multibase": "^0.7.0"
-  },
-  "gitHead": "1e3fb240a50a8128fe285c0ed444286d87c996b2"
+  }
 }
